printk: support simple impl based on legacy_putchar

+++ crates/linux_adaptor/kernel_modules/cl/adaptor.c
@@ -7,5 +7,5 @@ unsigned long linux_virt_to_phys(unsigned long va)
 
 unsigned long linux_phys_to_virt(unsigned long pa)
 {
-    return __va(pa);
+    return (unsigned long) __va(pa);
 }
